Moving projection creation to ProjectionFactory
