+++ Source/Kernel/Events.Projections/Json/JsonProjectionSerializer.cs
@@ -3,10 +3,8 @@
 
 using Cratis.Concepts;
 using Cratis.Events.Projections.Definitions;
-using Cratis.Events.Projections.Expressions;
 using Cratis.Properties;
 using Newtonsoft.Json;
-using Newtonsoft.Json.Schema;
 
 namespace Cratis.Events.Projections.Json
 {
@@ -15,17 +13,13 @@ namespace Cratis.Events.Projections.Json
     /// </summary>
     public class JsonProjectionSerializer : IJsonProjectionSerializer
     {
-        readonly IPropertyMapperExpressionResolvers _propertyMapperExpressionResolvers;
         readonly JsonSerializer _serializer;
 
         /// <summary>
         /// Initializes a new instance of the <see cref="JsonProjectionSerializer"/>.
         /// </summary>
-        /// <param name="propertyMapperExpressionResolvers"><see cref="IPropertyMapperExpressionResolvers"/> for resolving event value expressions.</param>
-        public JsonProjectionSerializer(IPropertyMapperExpressionResolvers propertyMapperExpressionResolvers)
+        public JsonProjectionSerializer()
         {
-            _propertyMapperExpressionResolvers = propertyMapperExpressionResolvers;
-
             _serializer = new JsonSerializer();
             _serializer.Converters.Add(new PropertyPathJsonConverter());
             _serializer.Converters.Add(new PropertyExpressionDictionaryJsonConverter());
@@ -44,62 +38,5 @@ namespace Cratis.Events.Projections.Json
 
         /// <inheritdoc/>
         public ProjectionDefinition Deserialize(string json) => _serializer.Deserialize<ProjectionDefinition>(new JsonTextReader(new StringReader(json)))!;
-
-        /// <inheritdoc/>
-        public IProjection CreateFrom(ProjectionDefinition definition) =>
-            CreateProjectionFrom(
-                definition,
-                PropertyPath.Root,
-                PropertyPath.Root,
-                $"Root({definition.Identifier})",
-                new Dictionary<PropertyPath, ChildrenDefinition>(),
-                _ => { });
-
-        IProjection CreateProjectionFrom(
-            ProjectionDefinition projectionDefinition,
-            PropertyPath childrenAccessorProperty,
-            PropertyPath identifiedByProperty,
-            ProjectionPath path,
-            IDictionary<PropertyPath, ChildrenDefinition> childrenDefinitions,
-            Action<IEnumerable<EventTypeWithKeyResolver>> addChildEventTypes)
-        {
-            var eventsForProjection = projectionDefinition.From.Select(kvp => new EventTypeWithKeyResolver(kvp.Key, string.IsNullOrEmpty(kvp.Value.ParentKey) ?
-                    EventValueProviders.FromEventSourceId :
-                    EventValueProviders.FromEventContent(kvp.Value.ParentKey!))).ToList();
-
-            var childProjections = projectionDefinition.Children.Select(kvp => CreateProjectionFrom(
-                    kvp.Value,
-                    kvp.Key,
-                    kvp.Value.IdentifiedBy,
-                    $"{path} -> ChildrenAt({kvp.Key.Path})",
-                    projectionDefinition.Children,
-                    _ =>
-                    {
-                        eventsForProjection.AddRange(_);
-                        addChildEventTypes(_);
-                    })).ToArray();
-
-            var model = new Model(projectionDefinition.Model.Name, JSchema.Parse(projectionDefinition.Model.Schema));
-            addChildEventTypes(eventsForProjection);
-
-            var projection = new Projection(projectionDefinition.Identifier, path, model, eventsForProjection, childProjections);
-
-            foreach (var (childrenProperty, childrenDefinition) in childrenDefinitions)
-            {
-                foreach (var (eventType, fromDefinition) in childrenDefinition.From)
-                {
-                    var propertyMappers = fromDefinition.Properties.Select(kvp => _propertyMapperExpressionResolvers.Resolve(kvp.Key, kvp.Value));
-                    projection.Event.From(eventType).Child(childrenProperty, childrenDefinition.IdentifiedBy, EventValueProviders.FromEventSourceId, propertyMappers);
-                }
-            }
-
-            foreach (var (eventType, fromDefinition) in projectionDefinition.From)
-            {
-                var propertyMappers = fromDefinition.Properties.Select(kvp => _propertyMapperExpressionResolvers.Resolve(kvp.Key, kvp.Value));
-                projection.Event.From(eventType).Project(childrenAccessorProperty, identifiedByProperty, EventValueProviders.FromEventSourceId, propertyMappers);
-            }
-
-            return projection;
-        }
     }
 }
